pyproject.toml
[project]
name = "HDP"
version = "0.1.0"
description = "Summarization and classification of abstracts with PG-LLaMA"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "datasets>=3.5.0",
    "fastapi[standard]>=0.115.12",
    "huggingface>=0.0.1",
    "huggingface-hub>=0.30.2",
    "loguru>=0.7.3",
    "xformers",
    "psutil>=7.0.0",
    "setuptools>=80.4.0",
    "python-dotenv>=1.1.0",
    "tabulate>=0.9.0",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
    "matplotlib>=3.10.3",
    "xgboost>=3.0.1",
    "evaluate>=0.4.3",
    "rouge-score>=0.1.2",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.41",
    "scrapy>=2.13.0",
    "openpyxl>=3.1.5",
    "spacy[cuda12x]>=3.8.6",
    "google-genai>=1.16.1",
    "transformers[torch]>=4.51",
    "unsloth>=2025.4.7",
    "unsloth-zoo>=2025.4.4",
    "torch>=2.7",
    "triton>=3.3.0",
    "codecarbon>=2.7.1",
    "tensorboardx>=2.6.2.2",
    "tensorboard>=2.19.0",
    "pip>=25.1.1",
    "wheel>=0.45.1",
    "spacy-llm>=0.7.3",
    "flair>=0.15.1",
    "tf-keras>=2.15.0",
]

[project.optional-dependencies]
build = ["setuptools"]

[tool.uv.sources]
torch = [{ index = "pytorch-cu128" }]
torchvision = [{ index = "pytorch-cu128" }]

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[dependency-groups]
dev = []
